requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
python-dotenv==1.0.1
redis==5.0.7
requests==2.31.0
soupsieve==2.5
spotipy==2.24.0
urllib3==2.1.0
